+++ rust/main/agents/relayer/src/msg/message_processor.rs
@@ -160,7 +160,10 @@ impl MessageProcessor {
 
         let entrypoint = self.payload_dispatcher_entrypoint.take().map(Arc::new);
 
-        let prepare_task = self.create_classic_prepare_task();
+        let prepare_task = match &entrypoint {
+            None => self.create_classic_prepare_task(),
+            Some(entrypoint) => self.create_lander_prepare_task(entrypoint.clone()),
+        };
 
         let submit_task = match &entrypoint {
             None => self.create_classic_submit_task(),
@@ -581,7 +584,7 @@ async fn submit_classic_task(
                 continue;
             }
             std::cmp::Ordering::Equal => {
-                let op = batch.pop().unwrap();
+                let op = batch.pop().expect("Should not happen");
                 submit_single_operation(op, &mut prepare_queue, &mut confirm_queue, &metrics).await;
             }
             std::cmp::Ordering::Greater => {