+++ solidity/contracts/token/interfaces/IFiatToken.sol
@@ -21,4 +21,11 @@ interface IFiatToken is IERC20 {
      * @return True if the operation was successful.
      */
     function mint(address _to, uint256 _amount) external returns (bool);
+
+    /**
+     * @notice Gets the minter allowance for an account.
+     * @param _minter The address to check.
+     * @return The remaining minter allowance for the account.
+     */
+    function minterAllowance(address _minter) external view returns (uint256);
 }