+++ typescript/cosmos-sdk/CHANGELOG.md
@@ -1,10 +1,43 @@
 # @hyperlane-xyz/cosmos-sdk
 
+<<<<<<< HEAD
+
 ## 15.0.0-beta.0
 
 ### Patch Changes
 
-- @hyperlane-xyz/cosmos-types@15.0.0-beta.0
+- # @hyperlane-xyz/cosmos-types@15.0.0-beta.0
+
+## 16.2.0
+
+### Patch Changes
+
+- @hyperlane-xyz/cosmos-types@16.2.0
+
+## 16.1.1
+
+### Patch Changes
+
+- @hyperlane-xyz/cosmos-types@16.1.1
+
+## 16.1.0
+
+### Patch Changes
+
+- @hyperlane-xyz/cosmos-types@16.1.0
+
+## 16.0.0
+
+### Patch Changes
+
+- @hyperlane-xyz/cosmos-types@16.0.0
+
+## 15.0.0
+
+### Patch Changes
+
+- @hyperlane-xyz/cosmos-types@15.0.0
+  > > > > > > > main
 
 ## 14.4.0